refactor: rimuovi re-export da cli/ dopo spostamento in domain/
Elimina 5 file di re-export (cli/layer_ops, catalog_ops, inspect/report_ops,
inspect/readiness_ops, preflight_ops). Rimuove re-export da cli/common.py.
Aggiorna test a importare direttamente da domain/.
cli/ resta solo con comandi Typer + dump_cfg_section + load_cfg_and_logger.
