+++ toolkit/cli/common.py
@@ -4,12 +4,6 @@ from typing import Any
 
 from toolkit.core.config import load_config
 from toolkit.core.logging import get_logger
-from toolkit.domain.common import (  # noqa: F401  # re-export
-    format_profile_preview,
-    iter_selected_years,
-    iter_years,
-    load_layer_profile_summaries,
-)
 
 
 def dump_cfg_section(cfg_section: Any) -> Any:
